remove other copies of plugin if more than one
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@4166 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ PluginManagerController.m
@@ -387,13 +387,19 @@ NSInteger sortPluginArrayByName(id plugin1, id plugin2, void *context)
 	
 	NSString *installDirectoryPath = [PluginManager userActivePluginsDirectoryPath]; // default = user active directory
 	
+	BOOL alreadyFound = NO;
+	
 	for (NSString *dir in directories) // search if the plugin was already installed
 	{
 		if([[NSFileManager defaultManager] fileExistsAtPath:[dir stringByAppendingPathComponent:[pluginPath lastPathComponent]]])
 		{
-			installDirectoryPath = dir; // in that case, install the (updated) plugin in the same directory it was
+			if(!alreadyFound)
+			{
+				installDirectoryPath = dir; // in that case, install the (updated) plugin in the same directory it was
+				alreadyFound = YES;
+			}
 			[[NSFileManager defaultManager] removeFileAtPath:[dir stringByAppendingPathComponent:[pluginPath lastPathComponent]] handler:nil];
-			break;
+			//break;
 		}
 	}
